docker/agent/app/requirements.txt
influxdb_client>=1.36.1
psutil>=5.9.5
python-dotenv>=1.0.0
